Add files via upload
A US-based housing company named Surprise Housing has decided to enter the Australian market. The company uses data analytics to purchase houses at a price below their actual values and flip them on at a higher price. For the same purpose, the company has collected a data set from the sale of houses in Australia. 
The company wants to know:

1.Which variables are significant in predicting the price of a house, and

2.How well those variables describe the price of a house.